fix(typecheck): tsc checks test/ too — the green was wider than it was
`tsconfig.json`'s include was `["src/**/*"]`, so every "tsc clean" this repo
ever reported covered the ~290 colocated src/**/__tests__ files and skipped all
127 under test/. Two days ago that let a `StampSpec.pri` widening leave three
call sites in test/art/composer-buffer.test.ts mistyped with tsc reporting
clean; a human found them by reading the file.

Blast radius, measured before fixing: 15 errors across 10 files, 5 error codes,
132 files newly entering the program.

ONE config, not a tsconfig.test.json: the test files need nothing looser. All
127 import describe/it/expect from 'vitest' explicitly, none use a path alias,
and the default lib for target ES2022 already carries DOM. Not one compiler
option was relaxed — the whole tsconfig diff is the include array and a comment.

No suppressions: zero @ts-expect-error, zero any, zero exclusions. Two
`as unknown as` double casts in tests that deliberately build off-contract
values, both newly commented with what they assert.

Real find: test/editing/set-collision-edit.test.ts built Section.collisionEdit
as a Uint8Array where it is a Uint16Array plane of 16-bit engine attr words. It
stayed green because 40 and 12 fit in a byte — a false green inside a green
test, from the same missing check.

Adds `npm run typecheck`; cite the script, not a hand-typed npx invocation.

Red-first, plant in test/: restoring `pri: false` at composer-buffer.test.ts:117
is silent under the old include (exit 0) and fails under the new script naming
file, line, column and type. Suite unchanged: 5637 passed / 7 skipped.

Packet: docs/reviews/2026-08-29-typecheck-tests.md

+++ test/editing/set-collision-edit.test.ts
@@ -5,7 +5,11 @@ import type { S4Level } from '../../src/core/editing/commands';
 
 function level(): S4Level {
   const s = createSection(0, 'S0');
-  s.collisionEdit = new Uint8Array(256 * 256);
+  // Uint16Array, not Uint8Array: `Section.collisionEdit` is a plane of 16-bit
+  // engine attribute words. This file used to build a Uint8Array — the writes
+  // it exercises happen to fit in a byte, so the suite stayed green while
+  // testing an array width production never hands the command.
+  s.collisionEdit = new Uint16Array(256 * 256);
   return { sections: [s] };
 }
 
@@ -13,7 +17,7 @@ describe('set-collision-edit', () => {
   it('applies and undoes attr writes on the chosen plane', () => {
     const h = new EditHistory();
     const lv = level();
-    lv.sections[0]!.collisionEditB = new Uint8Array(256 * 256);
+    lv.sections[0]!.collisionEditB = new Uint16Array(256 * 256);
     h.execute({ type: 'set-collision-edit', plane: 'a', description: 'paint', sectionIndex: 0,
       entries: [{ index: 5, oldColl: 0, newColl: 40 }] }, lv);
     h.execute({ type: 'set-collision-edit', plane: 'b', description: 'paint', sectionIndex: 0,
